refactor(form sections): move files into dedicated folder
