+++ src/pages/import/Data.js
@@ -10,12 +10,12 @@ import { DataIcon } from '../../components/Icon'
 import { DryRun } from '../../components/Inputs/DryRun'
 import { Error } from '../../components/Error'
 import { File } from '../../components/FinalFormComponents/File'
-import { FormContent } from '../../components/FormContent'
-import { FormFoot } from '../../components/FormFoot'
-import { FormHead } from '../../components/FormHead'
+import { FormContent } from '../../components/FormSections/FormContent'
+import { FormFoot } from '../../components/FormSections/FormFoot'
+import { FormHead } from '../../components/FormSections/FormHead'
+import { MoreOptions } from '../../components/FormSections/MoreOptions'
 import { Format } from '../../components/Inputs/Format'
 import { IdScheme } from '../../components/Inputs/idScheme'
-import { MoreOptions } from '../../components/MoreOptions'
 import { OrgUnitIdScheme } from '../../components/Inputs/OrgUnitIdScheme'
 import { PreheatCache } from '../../components/Inputs/PreheatCache'
 import { Progress } from '../../components/Loading/Progress'
